+++ src/tests/EntityGraphQL.Tests/SortTests/SortTests.cs
@@ -193,6 +193,36 @@ namespace EntityGraphQL.Tests
             Assert.Equal(1, Enumerable.ElementAt(project.tasks, 3).id);
         }
         [Fact]
+        public void SupportUseSortOnNonRootVariableWithClass()
+        {
+            var schema = SchemaBuilder.FromObject<TestDataContext>(false);
+            schema.Type<Project>().GetField("tasks", null)
+                .UseSort();
+            var gql = new QueryRequest
+            {
+                Query = @"query($sort: PersonSortInput) {
+                    projects {
+                        tasks(sort: $sort) { id }
+                    }
+                }",
+                Variables = new QueryVariables
+                {
+                    { "sort", new IdSort { Id = SortDirectionEnum.DESC } }
+                }
+            };
+            var context = new TestDataContext().FillWithTestData();
+            var tree = schema.ExecuteRequest(gql, context, null, null);
+            Assert.Null(tree.Errors);
+            dynamic projects = ((IDictionary<string, object>)tree.Data)["projects"];
+            Assert.Equal(1, Enumerable.Count(projects));
+            var project = Enumerable.First(projects);
+            Assert.Equal(4, Enumerable.Count(project.tasks));
+            Assert.Equal(4, Enumerable.ElementAt(project.tasks, 0).id);
+            Assert.Equal(3, Enumerable.ElementAt(project.tasks, 1).id);
+            Assert.Equal(2, Enumerable.ElementAt(project.tasks, 2).id);
+            Assert.Equal(1, Enumerable.ElementAt(project.tasks, 3).id);
+        }
+        [Fact]
         public void SupportUseSortOnNonRoot2()
         {
             var schema = SchemaBuilder.FromObject<TestDataContext>();
@@ -217,6 +247,11 @@ namespace EntityGraphQL.Tests
             Assert.Equal(2, Enumerable.ElementAt(project.tasks, 2).id);
             Assert.Equal(1, Enumerable.ElementAt(project.tasks, 3).id);
         }
+    }
+
+    internal class IdSort
+    {
+        public SortDirectionEnum Id { get; set; }
     }
 
     internal class TestArgs